+++ tests/testthat/test-loading.R
@@ -36,7 +36,7 @@ test_that("Retrieve path with libname", {
 
 test_that("Retrieve files from path with libname", {
     expect_message(my_path <- libname(external_path, get_files = TRUE), " > Filepaths successfully retrieved: ")
-    expect_equal(names(my_path), c("qol_example_data.csv", "qol_example_data.xlsx"))
+    expect_equal(names(my_path), c("qol_example_data_csv.csv", "qol_example_data_xlsx.xlsx"))
 })
 
 ###############################################################################